+++ sphinx/source/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 -------
 * Added a theme tune and an 'all shields' tune for Skool Daze Take Too
 * Added a theme tune and an 'all shields' tune for Ezad Looks
+* Added a theme tune and an 'all shields' tune for Back to Skool Daze
 
 1.1.1 (2014-01-26)
 ------------------